+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 - update packages ([bea31d2](https://github.com/air-supply94/zhouhaifei-react-scripts/commit/bea31d279a3d1b3f3315ca1ad3c2d7b00767d52e))
+- **react-scripts:** 移动端全部转 rem ([b170e6c](https://github.com/air-supply94/zhouhaifei-react-scripts/commit/b170e6cc4940c960a63fbb14ba5bf4145b947d57))
 
 ## [2.10.9](https://github.com/air-supply94/zhouhaifei-react-scripts/compare/v2.10.8...v2.10.9) (2021-04-20)