+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# [0.0.0](https://github.com/air-supply94/zhouhaifei-react-scripts/compare/v2.10.9...v0.0.0) (2021-04-20)
+# [0.0.0](https://github.com/air-supply94/zhouhaifei-react-scripts/compare/v2.10.10...v0.0.0) (2021-04-20)
+
+### Features
+
+- **react-scripts:** 内联 px2rem ([64828c8](https://github.com/air-supply94/zhouhaifei-react-scripts/commit/64828c8aea62910e83854f93d3decb888f49ea4a))
+
+## [2.10.10](https://github.com/air-supply94/zhouhaifei-react-scripts/compare/v2.10.9...v2.10.10) (2021-04-20)
 
 ### Features